requirements.txt
google-generativeai
chromadb
pydantic
python-dotenv
pyyaml
colorama
